requirements.txt
requests~=2.28.1
schedule~=1.1.0
certifi~=2022.6.15
paho-mqtt~=1.6.1